requirements.txt
openpyxl
flask==2.3.2
openai==0.27.8
pandas==2.0.3
python-dotenv
plotly==5.18.0
numpy==1.24.3
typing
Flask-Cors==4.0.0
scipy==1.11.3
matplotlib==3.8.1
scikit-learn==1.3.2
asgiref